requirements.txt
Django>=3.0.5,<3.1.0
djangorestframework>=3.11.0,<3.12.0
flake8>=3.7.9,<3.8.0
psycopg2>=2.8.5,<2.9.0
python-dateutil>=2.8.1,<2.9.0
Faker>=4.0.3,<4.1.0
